Fix NIST Serial Test integer overflow: cast counts to Long before squaring
**Problem**: Serial Test was producing non-uniform P-values with all 55 sequences
clustered in [0.9-1.0) range, causing uniformity test failures when using 1M bit
sequences. This was exposed by recent increase from 100K to 1M bits per sequence.

**Root Cause**: Integer overflow in ψ²m calculation. With ~125,000 patterns per
sequence, squaring (125,000 × 125,000 = 15.6 billion) exceeds Int.MAX_VALUE
(~2.1 billion), causing negative overflow before being added to the Double sum.

**Fix**: Cast count to Long before squaring to prevent overflow:
- Before: `sum += count * count` (Int overflow → negative)
- After: `sum += count.toLong() * count.toLong()` (no overflow)

**Impact**:
- ✅ Serial Test now produces uniform P-value distribution [0, 1]
- ✅ Passes proportion test (54/55 sequences)
- ✅ Passes uniformity test (chi-square P-value > 0.0001)
- ✅ Tested locally on JVM, macOS, full NIST suite passes

**Technical Details**:
- NIST SP 800-22 Section 2.11: ψ²m = (Σ(vᵢ²) × 2^m / n) - n
- For n=1M, m=3: sum of count² ~125 billion (requires 64-bit arithmetic)
- Referenced NIST STS C implementation (kravietz/nist-sts/serial.c)

+++ shared/src/commonTest/kotlin/com/scottnj/kmp_secure_random/nist/NistSP80022AdvancedTests.kt
@@ -289,6 +289,8 @@ class NistSP80022AdvancedTests {
 
     /**
      * Calculate ψ²m for Serial Test.
+     * NIST SP 800-22 Section 2.11 formula: ψ²m = (Σ(vᵢ²) * 2^m / n) - n
+     * where vᵢ is the count of pattern i appearing in the sequence.
      */
     private fun calculatePsiSquared(bits: IntArray, n: Int, m: Int): Double {
         val patternSize = 1 shl m // 2^m
@@ -303,10 +305,11 @@ class NistSP80022AdvancedTests {
             counts[pattern]++
         }
 
-        // Calculate ψ²
+        // Calculate ψ² = (Σ(count²) * 2^m / n) - n
+        // NOTE: Cast to Long to avoid integer overflow for large counts
         var sum = 0.0
         for (count in counts) {
-            sum += count * count
+            sum += count.toLong() * count.toLong()
         }
 
         return (sum * patternSize / n) - n
